+++ auctions/models.py
@@ -90,9 +90,6 @@ class Bid(models.Model):
     listing = models.ForeignKey(Listing, on_delete=models.DO_NOTHING)
     active = models.BooleanField(default=True)
 
-    def auction_winner(self, Listing):
-        self.listing = Listing.id
-
     def __str__(self):
         return (
             "Contact ID: " + str(self.owner_id) + "Listing ID: " + str(self.listing_id)
@@ -107,29 +104,6 @@ class Bid(models.Model):
         if current_bid is None:
             current_bid = self.listing.start_price
         return current_bid
-
-    # def validate_minimum_bid(self):
-    # bids must be greater than previous bids on the same listing
-
-    # if current_highest_bid["bid__max"] is None:
-    #     current_highest_bid["bid__max"] = self.listing.start_price
-    # else:
-    #     current_highest_bid["bid__max"] += 1
-    # if not self.bid >= current_highest_bid["bid__max"]:
-    #     raise ValidationError(
-    #         {"bid": f'The current minimum bid is {current_highest_bid["bid__max"] + 1.0}'}
-    #    )
-
-    def clean(self, *args, **kwargs):
-        # self.validate_negative_bid()
-        # TODO validate min bid is happening when bid_form.is_valid() is called
-        # TODO bind listinb object to bid form before is_valid() is called
-        # self.validate_minimum_bid()
-        super().clean(*args, **kwargs)
-
-    def save(self, *args, **kwargs):
-
-        return super(Bid, self).save(*args, **kwargs)
 
 
 class Comment(models.Model):
