+++ compose/local/django/celery/flower/start
@@ -5,7 +5,7 @@ set -o nounset
 
 
 celery \
-    -A commerce.celery_app \
+    -A tasks.celery_app \
     -b "${CELERY_BROKER_URL}" \
     flower \
     --basic_auth="${CELERY_FLOWER_USER}:${CELERY_FLOWER_PASSWORD}"
