+++ compose/local/django/celery/worker/start
@@ -4,4 +4,4 @@ set -o errexit
 set -o nounset
 
 
-watchgod celery.__main__.main --args -A commerce.celery_app worker -l INFO
+watchgod celery.__main__.main --args -A tasks.celery_app worker -l INFO
